Cambiada la clave primaria de User para que sea el username, cambiado los metodos del DAO para casar con la nueva estructura. Sigue sin revisarse las lineas 111 a 114, que podrían resumirse en u = user creo. Tanto al crear un nuevo usuario como al hacer un update se revisa si existe o da error.
Changed the primary key of User from the ID to the Username. Changed the methods of the DAO to fit in the new structure. Still unrevised lines 111 to 114, which could be changed to u = user (I think). Both when creating a new User and when updating an already existing one the code now checks if its exist in the database for different purposes (so it gaves an error when you cannot proceed with the creation/update).